pyproject.toml
[project]
name = "geolocateai"
version = "0.1.0"
description = ""
authors = [
    {name = "Pop101"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars (>=1.27.1,<2.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "hdbscan (>=0.8.40,<0.9.0)",
    "lxml (>=5.3.2,<6.0.0)",
    "transformers (>=4.51.3,<5.0.0)",
]

[tool.poetry.group.dev.dependencies]
pandas = ">=2.2.3,<3.0.0"
pyarrow = ">=20.0.0,<21.0.0"
jupyter = ">=1.1.1,<2.0.0"
seaborn = "^0.13.2"
datashader = ">=0.18.0,<0.19.0"
msgpack = ">=1.1.0,<2.0.0"
piexif = ">=1.1.3,<2.0.0"
geopy = ">=2.4.1,<3.0.0"
colorcet = ">=3.1.0,<4.0.0"
holoviews = ">=1.20.2,<2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
